Add unzip command and gui
This commit adds a new command to unzip a zipped message. The command
can be triggered either via the GUI or through the command palette.

The command works by hooking into a new namespace 'servicebusmessagegzip'
and mapping a provider to that namespace. If a document is loaded from
this namespace, the gzipMessageProvider will try to unzip the message
using 'zlib'.

Ref DigitalMolecules/vscode-servicebus-explorer#44